.travis.yml: Move unsupported check to a script
The unsupported check was previously a hack inside the main
Travis CI `script` block, and was not correct.
Travis CI now supports custom `script` blocks for each build
matrix entry, which allows the logic to be moved to an
external and more clear script.
This also allows the invocation of codecov to be in the
standard `after_success` and `after_failure` blocks.

Related to https://github.com/coala/coala/issues/4372
Related to https://github.com/coala/coala/issues/4323