FileFactory: Add missing magic methods
`__len__` and `__getitem__` are missing magic methods of
dict needed to be emulated in order to be backwards compatible.

`__len__` is a reasonable action for any container, including
container proxies, however `__getitem__` is deprecated as
it requires efficient lookup, which is an IO issue for files,
or is a memory issue for large codebases if all files are
kept in memory.

These are needed by coala-quickstart green mode to work on master.

Closes https://github.com/coala/coala/issues/5810