Linter: Avoid processing empty streams
When a linter returns an empty stream, it should not be
processed.  In the case of process_output_corrected,
processing it is extremely destructive, zero-ing the
file.  In the other modes, there is no output to be
processed, so it is merely wasted processing.

Also emit warnings for any messages found on a
stream which is not being processed.

Followup enhancements might add ways for linter bears
to skip known messages where a linter is noisy.

Fixes https://github.com/coala/coala/issues/5415